requirements.txt
diffusers
sentencepiece
protobuf
transformers
accelerate
Pillow
opencv-python
compel
xformers<0.0.27
trl<0.9.0 
peft
bitsandbytes
lpips
scikit-image
pytorch-msssim
omegaconf
einops
imageio